refactor: move repository classes to domain packages
