+++ src/main/java/kitchenpos/common/infra/DefaultPurgomalumClient.java
@@ -7,6 +7,8 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
 
+import kitchenpos.common.domain.PurgomalumClient;
+
 @Component
 public class DefaultPurgomalumClient implements PurgomalumClient {
     private final RestTemplate restTemplate;
