+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -15,12 +15,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.infra.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.infra.MenuRepository;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.infra.ProductRepository;
-import kitchenpos.common.infra.PurgomalumClient;
+import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.common.domain.PurgomalumClient;
 import kitchenpos.menu.application.dto.MenuCreationRequest;
 
 @Service
