+++ src/main/java/kitchenpos/menu/domain/MenuName.java
@@ -6,7 +6,7 @@ import org.springframework.util.StringUtils;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import kitchenpos.common.infra.PurgomalumClient;
+import kitchenpos.common.domain.PurgomalumClient;
 
 @Embeddable
 public class MenuName {
