+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.infra;
+package kitchenpos.menu.domain;
 
 import java.util.List;
 import java.util.Optional;
