+++ src/main/java/kitchenpos/order/infra/JpaOrderTableRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.UUID;
 
 import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 
 public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
 }
