+++ src/main/java/kitchenpos/product/domain/ProductName.java
@@ -6,7 +6,7 @@ import org.apache.logging.log4j.util.Strings;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import kitchenpos.common.infra.PurgomalumClient;
+import kitchenpos.common.domain.PurgomalumClient;
 
 @Embeddable
 public class ProductName {
