+++ src/main/java/kitchenpos/product/infra/JpaProductRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.UUID;
 
 import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 
 public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
 }
