+++ src/test/java/kitchenpos/Fixtures.java
@@ -12,7 +12,7 @@ import kitchenpos.order.domain.OrderTable;
 import kitchenpos.order.domain.OrderType;
 import kitchenpos.product.domain.Product;
 import kitchenpos.order.domain.TakeoutOrder;
-import kitchenpos.common.infra.PurgomalumClient;
+import kitchenpos.common.domain.PurgomalumClient;
 import kitchenpos.order.application.dto.OrderCreationRequest;
 import kitchenpos.order.application.dto.OrderLineItemCreationRequest;
