+++ src/test/java/kitchenpos/menu/domain/MenuTest.java
@@ -17,7 +17,7 @@ import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import kitchenpos.common.infra.FakePurgomalumClient;
-import kitchenpos.common.infra.PurgomalumClient;
+import kitchenpos.common.domain.PurgomalumClient;
 import kitchenpos.menu.application.dto.MenuCreationRequest;
 import kitchenpos.product.domain.Product;
