+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -25,16 +25,16 @@ import kitchenpos.order.factory.OrderFactoryProvider;
 import kitchenpos.order.domain.DeliveryOrder;
 import kitchenpos.order.domain.EatInOrder;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.infra.MenuRepository;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.infra.InMemoryOrderFactoryProvider;
 import kitchenpos.order.infra.InMemoryOrderRepository;
 import kitchenpos.order.infra.InMemoryOrderTableRepository;
-import kitchenpos.order.infra.OrderRepository;
+import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.infra.OrderTableRepository;
+import kitchenpos.order.domain.OrderTableRepository;
 import kitchenpos.order.domain.OrderType;
 import kitchenpos.order.application.dto.OrderCreationRequest;
 import kitchenpos.order.application.dto.OrderLineItemCreationRequest;
