+++ src/test/java/kitchenpos/order/infra/InMemoryOrderRepository.java
@@ -2,7 +2,7 @@ package kitchenpos.order.infra;
 
 import kitchenpos.order.domain.EatInOrder;
 import kitchenpos.order.domain.Order;
-import kitchenpos.order.infra.OrderRepository;
+import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderTable;
 import kitchenpos.order.domain.OrderType;
