+++ src/test/java/kitchenpos/order/infra/InMemoryOrderTableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.infra;
 
 import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.infra.OrderTableRepository;
+import kitchenpos.order.domain.OrderTableRepository;
 
 import java.util.ArrayList;
 import java.util.HashMap;
