+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -4,10 +4,10 @@ import kitchenpos.common.infra.FakePurgomalumClient;
 import kitchenpos.product.infra.InMemoryProductRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.infra.InMemoryMenuRepository;
-import kitchenpos.menu.infra.MenuRepository;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.infra.ProductRepository;
-import kitchenpos.common.infra.PurgomalumClient;
+import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.common.domain.PurgomalumClient;
 import kitchenpos.product.application.dto.ProductCreationRequest;
 
 import org.junit.jupiter.api.BeforeEach;
