+++ src/test/java/kitchenpos/product/domain/ProductTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import kitchenpos.common.infra.FakePurgomalumClient;
-import kitchenpos.common.infra.PurgomalumClient;
+import kitchenpos.common.domain.PurgomalumClient;
 
 class ProductTest {
